apps/requirements.txt
Django==3.1.5
gunicorn==20.0.4
djangorestframework
pillow
django-mathfilters
django-modeltranslation
djangorestframework_simplejwt